including cil visitors objects in the build_cil_ast logic, renaming parser.py to cparser.py (for cool parser) due to conflicts in python3.8
